+++ react/__mocks__/vtex.styleguide.tsx
@@ -1,3 +1,4 @@
 export { default as Card } from '@vtex/styleguide/lib/Card'
 export { default as Input } from '@vtex/styleguide/lib/Input'
 export { default as Button } from '@vtex/styleguide/lib/Button'
+export { default as Box } from '@vtex/styleguide/lib/Box'