+++ react/typings/@vtex/styleguide.d.ts
@@ -1,3 +1,4 @@
 declare module '@vtex/styleguide/lib/Card'
 declare module '@vtex/styleguide/lib/Input'
 declare module '@vtex/styleguide/lib/Button'
+declare module '@vtex/styleguide/lib/Box'