+++ react/__mocks__/vtex.pixel-manager.ts
@@ -1,0 +1,3 @@
+export const usePixel = () => {
+  return { push: jest.fn() }
+}