+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Feedback when the sla list is empty after the query
+
+### Added
+
 - Tests for the pickupSelector container
 - Tests for the zip code input